requirements.txt
# Core scientific computing
numpy>=2.3.0

# Plotting and visualization
matplotlib>=3.10.0

# Scientific computing
scipy>=1.16.0

# Data validation and settings management
pydantic>=2.11.0

# YAML parsing
ruamel.yaml>=0.18.0

# Network/graph visualization
networkx>=3.5.0